remove ONNX_INITIAL_TYPES
