+++ tests/test_onnxml_lightgbm_converter.py
@@ -104,33 +104,6 @@ class TestONNXLightGBMConverter(unittest.TestCase):
 
         np.testing.assert_allclose(onnx_ml_pred[0].flatten(), pt_model.predict(X))
 
-    # Check converter with extra configs.
-    @unittest.skipIf(
-        not (onnx_ml_tools_installed() and onnx_runtime_installed()), reason="ONNXML test require ONNX, ORT and ONNXMLTOOLS"
-    )
-    @unittest.skipIf(not lightgbm_installed(), reason="LightGBM test requires LightGBM installed")
-    def test_lightgbm_pytorch_extra_config(self):
-        warnings.filterwarnings("ignore")
-        X = [[0, 1], [1, 1], [2, 0]]
-        X = np.array(X, dtype=np.float32)
-        y = np.array([100, -10, 50], dtype=np.float32)
-        model = lgb.LGBMRegressor(n_estimators=3, min_child_samples=1)
-        model.fit(X, y)
-
-        # Create ONNX-ML model
-        onnx_ml_model = convert_lightgbm(
-            model, initial_types=[("input", FloatTensorType([X.shape[0], X.shape[1]]))], target_opset=9
-        )
-
-        # Create ONNX model
-        model_name = "hummingbird.ml.test.lightgbm"
-        extra_config = {}
-        extra_config[constants.ONNX_OUTPUT_MODEL_NAME] = model_name
-        extra_config[constants.ONNX_INITIAL_TYPES] = [("input", FloatTensorType([X.shape[0], X.shape[1]]))]
-        onnx_model = convert(onnx_ml_model, "onnx", extra_config=extra_config)
-
-        assert onnx_model.model.graph.name == model_name
-
     # Basic regression test.
     @unittest.skipIf(
         not (onnx_ml_tools_installed() and onnx_runtime_installed()), reason="ONNXML test require ONNX, ORT and ONNXMLTOOLS"
